+++ src/test/java/ru/yandex/practicum/filmorate/controller/ShouldCreateUsersWithoutNameInUserControllerTest2.java
@@ -1,0 +1,47 @@
+package ru.yandex.practicum.filmorate.controller;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.time.LocalDate;
+import java.util.List;
+
+import static org.junit.Assert.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+class ShouldCreateUsersWithoutNameInUserControllerTest2 {
+    @Autowired
+    private UserController userController;
+    User user;
+
+    @Test
+    public void shouldCreateUserWithoutName() {
+        user = User.builder()
+                .login("dolore")
+                .name("Nick Name")
+                .email("mail@mail.ru")
+                .birthday(LocalDate.of(1946, 8, 20))
+                .build();
+        user.setName("");
+        userController.addUser(user);
+
+        List<User> allUsers = userController.getAllUser();
+
+        assertEquals(1, allUsers.size(),
+                "Size Equal Test");
+
+        User userToCheck = allUsers.get(0);
+
+        assertEquals(user.getLogin(), userToCheck.getName(), "Login Equal Name Test");
+    }
+
+
+}
