+++ src/test/java/ru/yandex/practicum/filmorate/controller/ShouldGetAllFilmTestInFilmrControllerTest4.java
@@ -1,0 +1,42 @@
+package ru.yandex.practicum.filmorate.controller;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.time.LocalDate;
+import java.util.List;
+
+import static org.junit.Assert.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+class ShouldGetAllFilmTestInFilmrControllerTest4 {
+    @Autowired
+    private FilmController filmController;
+    Film film;
+
+    @Test
+    public void shouldGetAllFilmTest() {
+        film = Film.builder()
+                .name("nisi eiusmod")
+                .description("adipisicing")
+                .releaseDate(LocalDate.of(1967, 3, 25))
+                .duration(100)
+                .build();
+        filmController.addFilm(film);
+        Film film100 = film;
+        film100.setId(100);
+
+        filmController.addFilm(film100);
+
+        assertEquals(2, filmController.getAllFIlms().size(),
+                "Size Equal Test");
+    }
+}
